pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scriptcraft-python"
dynamic = ["version"]
description = "Data processing and quality control tools for research workflows"
readme = "README.md"
# license = "MIT"  # Temporarily removed to fix metadata issue
authors = [
    {name = "ScriptCraft Team", email = "scriptcraft@example.com"}
]
maintainers = [
    {name = "ScriptCraft Team", email = "scriptcraft@example.com"}
]
keywords = ["data-processing", "quality-control", "research", "validation", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "openpyxl>=3.0.0",
    "xlrd>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=12.0.0",
    "pydantic>=1.9.0",
    "python-dateutil>=2.8.0",
    "pytz>=2021.1",
    "python-docx>=0.8.11",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
web = ["selenium>=4.0.0"]
all = ["selenium>=4.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/mcusac/ScriptCraft-Workspace"
Documentation = "https://github.com/mcusac/ScriptCraft-Workspace#readme"
Repository = "https://github.com/mcusac/ScriptCraft-Workspace"
Issues = "https://github.com/mcusac/ScriptCraft-Workspace/issues"

[project.scripts]
scriptcraft = "scriptcraft.layers.layer_1_pypi.common.cli:main"
rhq-autofiller = "scriptcraft.layers.layer_1_pypi.tools.rhq_form_autofiller.main:main"
data-comparer = "scriptcraft.layers.layer_1_pypi.tools.data_content_comparer.main:main"
auto-labeler = "scriptcraft.layers.layer_1_pypi.tools.automated_labeler.main:main"
function-auditor = "scriptcraft.layers.layer_1_pypi.tools.function_auditor.main:main"
scriptcraft-release = "scriptcraft.layers.layer_1_pypi.cli.release_cli:main"

[tool.setuptools.dynamic]
version = {attr = "scriptcraft._version.__version__"}

[tool.setuptools.packages.find]
where = ["."]
include = ["scriptcraft*"]
exclude = ["tests*", "build*", "dist*", "*.egg-info*"]

[tool.setuptools.package-data]
"scriptcraft" = ["*.yaml", "*.yml", "*.json", "templates/**/*"]
"*" = ["*.txt", "*.md", "*.yaml", "*.yml"]

[tool.setuptools]
zip-safe = false 